chore: Refactor test code
