+++ crates/runtime/src/embeddings/index/query_table.rs
@@ -564,320 +564,19 @@ fn metadata_columns_to_exprs(metadata_columns: &MetadataColumns) -> Vec<Expr> {
 #[cfg(test)]
 mod tests {
 
-    use std::{any::Any, collections::HashMap, sync::Arc};
+    use std::{collections::HashMap, sync::Arc};
 
-    use arrow::{
-        array::{
-            ArrayData, ArrayRef, BooleanArray, FixedSizeListArray, Float32Array, Float64Array,
-            Int8Array, Int16Array, Int32Array, Int64Array, RecordBatch, StringArray, UInt8Array,
-            UInt16Array, UInt32Array, UInt64Array, new_null_array,
-        },
-        buffer::Buffer,
-        util::pretty,
-    };
-    use arrow_schema::{DataType, Field, Schema, SchemaRef};
-    use data_components::s3_vectors::{
-        MetadataColumn, MetadataColumns, S3_VECTOR_EMBEDDING_NAME, S3_VECTOR_PRIMARY_KEY_NAME,
-        query_provider::S3_VECTOR_DISTANCE_NAME,
-    };
+    use arrow_schema::{DataType, Field, Schema};
+    use data_components::s3_vectors::{S3_VECTOR_EMBEDDING_NAME, S3_VECTOR_PRIMARY_KEY_NAME};
     use datafusion::{
-        catalog::{MemTable, Session, TableProvider},
-        datasource::TableType,
-        error::DataFusionError,
-        logical_expr::TableProviderFilterPushDown,
-        physical_plan::{DisplayAs, ExecutionPlan},
-        prelude::{Expr, SessionConfig, SessionContext},
+        catalog::{MemTable, TableProvider},
         sql::TableReference,
     };
-    use search::generation::util::append_fields;
-    use snafu::ResultExt;
 
-    use crate::embeddings::index::{VectorIndex, VectorQueryTableProvider};
-
-    /// This is just a [`MemTable`] that pretends it can support all filter pushdowns.
-    /// This is useful for testing explain plans.
-    #[derive(Debug)]
-    pub struct ExplainMemTable(MemTable);
-
-    /// Wraps a [`ExecutionPlan`] with a new [`DisplayAs`] to show what filters have been pushed down.
-    /// This is useful for testing explain plans.
-    #[derive(Debug)]
-    pub struct ExplainExecutionPlan(
-        Arc<dyn ExecutionPlan>,
-        Vec<Expr>,
-        Option<usize>,
-        Option<Vec<usize>>,
-    );
-
-    impl ExecutionPlan for ExplainExecutionPlan {
-        fn name(&self) -> &'static str {
-            "ExplainExecutionPlan"
-        }
-
-        fn as_any(&self) -> &dyn Any {
-            self
-        }
-
-        fn properties(&self) -> &datafusion::physical_plan::PlanProperties {
-            self.0.properties()
-        }
-
-        fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-            self.0.children()
-        }
-
-        fn with_new_children(
-            self: Arc<Self>,
-            children: Vec<Arc<dyn ExecutionPlan>>,
-        ) -> datafusion::error::Result<Arc<dyn ExecutionPlan>> {
-            Ok(Arc::new(ExplainExecutionPlan(
-                Arc::clone(&self.0).with_new_children(children)?,
-                self.1.clone(),
-                self.2,
-                self.3.clone(),
-            )))
-        }
-
-        fn execute(
-            &self,
-            partition: usize,
-            context: Arc<datafusion::execution::TaskContext>,
-        ) -> datafusion::error::Result<datafusion::execution::SendableRecordBatchStream> {
-            self.0.execute(partition, context)
-        }
-    }
-
-    impl DisplayAs for ExplainExecutionPlan {
-        fn fmt_as(
-            &self,
-            _t: datafusion::physical_plan::DisplayFormatType,
-            f: &mut std::fmt::Formatter,
-        ) -> std::fmt::Result {
-            let columns: Vec<String> = self
-                .schema()
-                .fields()
-                .iter()
-                .map(|f| f.name())
-                .cloned()
-                .collect();
-
-            write!(
-                f,
-                "ExplainExecutionPlan: projection={columns:?} filter={:?} limit={:?}",
-                self.1, self.2,
-            )?;
-            Ok(())
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl TableProvider for ExplainMemTable {
-        fn as_any(&self) -> &dyn Any {
-            self
-        }
-
-        fn schema(&self) -> SchemaRef {
-            self.0.schema()
-        }
-
-        fn table_type(&self) -> TableType {
-            self.0.table_type()
-        }
-
-        async fn scan(
-            &self,
-            state: &dyn Session,
-            projection: Option<&Vec<usize>>,
-            filters: &[Expr],
-            limit: Option<usize>,
-        ) -> Result<Arc<dyn ExecutionPlan>, DataFusionError> {
-            Ok(Arc::new(ExplainExecutionPlan(
-                self.0.scan(state, projection, filters, limit).await?,
-                filters.to_vec(),
-                limit,
-                projection.cloned(),
-            )) as Arc<dyn ExecutionPlan>)
-        }
-
-        fn supports_filters_pushdown(
-            &self,
-            filters: &[&Expr],
-        ) -> Result<Vec<TableProviderFilterPushDown>, DataFusionError> {
-            Ok(vec![TableProviderFilterPushDown::Exact; filters.len()])
-        }
-    }
-
-    /// An implementation of [`VectorIndex`] that has one row. Useful for testing explain plans.
-    #[derive(Debug)]
-    pub struct PretendVectorIndex {
-        embedded_column: String,
-        primary_columns: Vec<Field>,
-        schema: Schema,
-        metadata: MetadataColumns,
-    }
-    impl PretendVectorIndex {
-        pub fn new(embedded_column: String, primary_columns: Vec<Field>, schema: Schema) -> Self {
-            let cols = schema
-                .fields()
-                .iter()
-                .filter_map(|f| {
-                    if f.name() == S3_VECTOR_PRIMARY_KEY_NAME
-                        || f.name() == S3_VECTOR_EMBEDDING_NAME
-                    {
-                        return None;
-                    }
-                    if f.metadata().get("filterable") == Some(&"true".to_string()) {
-                        Some(MetadataColumn::Filterable(Arc::clone(f)))
-                    } else {
-                        Some(MetadataColumn::NonFilterable(Arc::clone(f)))
-                    }
-                })
-                .collect::<Vec<_>>();
-
-            Self {
-                embedded_column,
-                primary_columns,
-                schema,
-                metadata: MetadataColumns::from(cols),
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl VectorIndex for PretendVectorIndex {
-        fn embedded_column(&self) -> String {
-            self.embedded_column.clone()
-        }
-
-        fn primary_fields(&self) -> Vec<Field> {
-            self.primary_columns.clone()
-        }
-
-        fn list_table_provider(&self) -> Arc<dyn TableProvider> {
-            Arc::new(ExplainMemTable(
-                MemTable::try_new(
-                    Arc::new(self.schema.clone()),
-                    vec![vec![one_row_default_record_batch_for_schema(&Arc::new(
-                        self.schema.clone(),
-                    ))]],
-                )
-                .expect("Could not build PretendVectorIndex::list_table_provider"),
-            ))
-        }
-
-        fn metadata_columns(&self) -> &MetadataColumns {
-            &self.metadata
-        }
-
-        fn augment_table(self: Arc<Self>, table: Arc<dyn TableProvider>) -> Arc<dyn TableProvider> {
-            table
-        }
-
-        async fn write(&self, _record: &RecordBatch) {}
-        async fn query_table_provider(
-            &self,
-            _query: &str,
-        ) -> Result<Arc<dyn TableProvider>, Box<dyn std::error::Error + Send + Sync>> {
-            let schema = append_fields(
-                &Arc::new(self.schema.clone()),
-                vec![Arc::new(Field::new(
-                    S3_VECTOR_DISTANCE_NAME,
-                    DataType::Float64,
-                    false,
-                ))],
-            );
-            println!("In query_table_provider schema={:?}", schema);
-            Ok(Arc::new(ExplainMemTable(
-                MemTable::try_new(
-                    Arc::clone(&schema),
-                    vec![vec![one_row_default_record_batch_for_schema(&schema)]],
-                )
-                .boxed()?,
-            )) as Arc<dyn TableProvider>)
-        }
-    }
-
-    pub async fn test_explain(
-        provider: Arc<dyn TableProvider>,
-        tbl: TableReference,
-        sql: &str,
-        snapshot_name: &str,
-    ) -> Result<(), String> {
-        let session =
-            SessionContext::new_with_config(SessionConfig::new().with_target_partitions(3));
-        session
-            .register_table(tbl, provider)
-            .map_err(|e| e.to_string())?;
-
-        let df = session
-            .sql(format!("EXPLAIN {sql}").as_str())
-            .await
-            .map_err(|e| e.to_string())?;
-
-        let col = df.collect().await.map_err(|e| e.to_string())?;
-        insta::assert_snapshot!(
-            snapshot_name,
-            format!(
-                "{}",
-                pretty::pretty_format_batches(&col).map_err(|e| e.to_string())?
-            )
-        );
-        Ok(())
-    }
-
-    #[allow(clippy::cast_sign_loss, clippy::cast_precision_loss)]
-    fn default_value_array(dt: &DataType) -> ArrayRef {
-        match dt {
-            DataType::Int8 => Arc::new(Int8Array::from(vec![0])) as ArrayRef,
-            DataType::Int16 => Arc::new(Int16Array::from(vec![0])) as ArrayRef,
-            DataType::Int32 => Arc::new(Int32Array::from(vec![0])) as ArrayRef,
-            DataType::Int64 => Arc::new(Int64Array::from(vec![0])) as ArrayRef,
-            DataType::UInt8 => Arc::new(UInt8Array::from(vec![0])) as ArrayRef,
-            DataType::UInt16 => Arc::new(UInt16Array::from(vec![0])) as ArrayRef,
-            DataType::UInt32 => Arc::new(UInt32Array::from(vec![0])) as ArrayRef,
-            DataType::UInt64 => Arc::new(UInt64Array::from(vec![0])) as ArrayRef,
-            DataType::Float32 => Arc::new(Float32Array::from(vec![0.0])) as ArrayRef,
-            DataType::Float64 => Arc::new(Float64Array::from(vec![0.0])) as ArrayRef,
-            DataType::Boolean => Arc::new(BooleanArray::from(vec![false])) as ArrayRef,
-            DataType::Utf8 => Arc::new(StringArray::from(vec![""])) as ArrayRef,
-            DataType::FixedSizeList(_, length) => {
-                let list_data_type = DataType::FixedSizeList(
-                    Arc::new(Field::new_list_field(DataType::Float32, false)),
-                    *length,
-                );
-                Arc::new(FixedSizeListArray::from(
-                    ArrayData::builder(list_data_type.clone())
-                        .len(1)
-                        .add_child_data(
-                            ArrayData::builder(DataType::Float32)
-                                .len(*length as usize)
-                                .add_buffer(Buffer::from_slice_ref(
-                                    (0..(*length as usize))
-                                        .map(|s| s as f32)
-                                        .collect::<Vec<_>>(),
-                                ))
-                                .build()
-                                .expect("unable to build FixedSizeListArray's ArrayData"),
-                        )
-                        .build()
-                        .expect("unable to build FixedSizeListArray"),
-                ))
-            }
-            _ => new_null_array(dt, 1),
-        }
-    }
-
-    /// Creates a [`RecordBatch`] with a single row that has default value of types, as per the [`Schema`].
-    fn one_row_default_record_batch_for_schema(schema: &Arc<Schema>) -> RecordBatch {
-        let arrays: Vec<ArrayRef> = schema
-            .fields()
-            .iter()
-            .map(|field| default_value_array(field.data_type()))
-            .collect();
-
-        RecordBatch::try_new(Arc::clone(schema), arrays)
-            .expect("could not build RecordBatch with one row")
-    }
+    use crate::embeddings::index::VectorQueryTableProvider;
+    use crate::embeddings::index::tests::{
+        PretendVectorIndex, one_row_default_record_batch_for_schema, test_explain,
+    };
 
     #[tokio::test]
     pub async fn test_vector_query_basic() -> Result<(), String> {
